feat(teleop): add the nav map as a PiP view
Extract the map page's canvas (occupancy grid + pose + plan + click-to-
navigate) into a reusable createMap() widget, and embed it in the camera
strip as a peer view with the same off / live-thumbnail / big ladder.

When promoted, the map covers the stage and the video stage hides; a
camera stays the session primary underneath — or, with all cameras off,
the WebRTC peer is released entirely (a map-only view costs zero
streaming, rebuilt on the next camera enable).

The teleop map is robot-centred so it stays legible when small, with
scroll-to-zoom whose level is saved separately for the thumbnail vs the
full-stage size. The standalone /map page keeps its fit-the-whole-grid view.
